+++ server/model/iosScheduledUpdateLog.go
@@ -12,7 +12,7 @@ const (
 
 // IOSScheduledUpdateLog logs the last time a device was updated.
 type IOSScheduledUpdateLog struct {
-	ID        uint32    `gorm:"primary_key;auto_increment;not_null" json:"id"`
+	ID        int64     `gorm:"primary_key;auto_increment;not_null" json:"id"`
 	DeviceID  string    `gorm:"index:idx_scheduled_update_log_device,unique" json:"deviceId"`
 	Device    IOSDevice `gorm:"constraint:OnDelete:CASCADE;" json:"device"`
 	Type      string    `gorm:"type:enum ('grades');" json:"type"`
